migrated players to next 14
