+++ package.json
@@ -57,6 +57,7 @@
     "husky": "^8.0.3",
     "lint-staged": "^15.1.0",
     "prettier": "^3.1.0",
+    "purify-css": "^1.2.5",
     "sass": "^1.69.5",
     "typescript": "^5"
   },
