+++ src/components/events/details.tsx
@@ -1,48 +1,28 @@
 import { countries } from "country-data"
 import Image from "next/image"
 import Link from "next/link"
-import openTabSection from "../../libs/tabs"
 import {
-  ComponentEventsSponsorship,
-  ComponentEventsTimetable,
   Enum_Event_Status,
   Event,
   EventLocation,
-  Maybe,
-  PlayerEntity,
   UploadFile,
   Venue,
 } from "../../models/graphql"
-import Gallery from "../layout/gallery"
 import Map from "../map"
-import PlayerGrid from "../players/grid"
 import EventDate from "./date"
-import EventDescription from "./description"
 import EventsNavigator from "./detailsnav"
-import EventSchedule from "./schedule"
 import EventSidebar from "./sidebar"
-import EventSponsorships from "./sponsorships"
 import UpcomingEventTimer from "./timer"
 //@ts-ignore
 import ReactCountryFlag from "react-country-flag"
-import HtmlContent from "../layout/html-content"
 import ICalendar from "./ical"
+import EventTabs from "./tabs"
 
-const EventDetails = (props: { event: Event }) => {
-  const { event } = props
+export default function EventDetails({ event }: { event: Event }) {
   const defaultImage = event.defaultImage.data?.attributes as UploadFile
   const eventLocation = event.location?.data?.attributes as EventLocation
   const venue = event.venue?.data?.attributes as Venue
-  const timetable = event.timetable as Array<Maybe<ComponentEventsTimetable>>
-  const players = event.players?.data as PlayerEntity[]
-  const hosts = event.hosts?.data as PlayerEntity[]
-  const mentors = event.mentors?.data as PlayerEntity[]
   const country = countries[eventLocation.country!].name
-
-  const participants = concatWithoutDuplicates(
-    players,
-    concatWithoutDuplicates(hosts, mentors),
-  )
 
   return (
     <>
@@ -181,89 +161,7 @@ const EventDetails = (props: { event: Event }) => {
           </div>
         </div>
 
-        {/* tabs */}
-        <div className="row">
-          <div className="courses-details-desc">
-            <ul className="nav nav-tabs" id="myTab" role="tablist">
-              {/* Overview */}
-              <li
-                className="current"
-                onClick={(e) => openTabSection(e, "overviewTab")}
-              >
-                Overview
-              </li>
-
-              {/* Registration */}
-              {event.status == Enum_Event_Status.Open &&
-                event.registration &&
-                event.registration.widgetCode && (
-                  <li onClick={(e) => openTabSection(e, "registrationTab")}>
-                    Registration
-                  </li>
-                )}
-
-              {/* Schedule */}
-              <li onClick={(e) => openTabSection(e, "scheduleTab")}>
-                Schedule
-              </li>
-
-              {/* Players */}
-              <li onClick={(e) => openTabSection(e, "playersTab")}>
-                Players{" "}
-                {participants && participants.length > 0
-                  ? `(${participants.length})`
-                  : ""}
-              </li>
-
-              {/* Photos */}
-              <li onClick={(e) => openTabSection(e, "photosTab")}>
-                Photos{" "}
-                {event.images && event.images.data.length > 0
-                  ? `(${event.images.data.length})`
-                  : ""}
-              </li>
-            </ul>
-
-            <div className="tab-content">
-              {/* Overview */}
-              <div id="overviewTab" className="tab-pane tabs_item">
-                {event.description && (
-                  <EventDescription description={event.description} />
-                )}
-                {hosts && <PlayerGrid title="Team" players={hosts} />}
-                {mentors && <PlayerGrid title="Mentors" players={mentors} />}
-                {event.sponsorships && (
-                  <EventSponsorships
-                    sponsorships={
-                      event.sponsorships as Array<ComponentEventsSponsorship>
-                    }
-                  />
-                )}
-              </div>
-              {/* Schedule */}
-              <div id="scheduleTab" className="tab-pane tabs_item">
-                {timetable && <EventSchedule timetable={timetable} />}
-              </div>
-
-              {/* Players */}
-              <div id="playersTab" className="tab-pane tabs_item">
-                {players && (
-                  <PlayerGrid title="Players" players={participants} />
-                )}
-              </div>
-
-              {/* Photos */}
-              <div id="photosTab" className="tab-pane tabs_item">
-                {event.images && <Gallery images={event.images.data} />}
-              </div>
-
-              {/* Registration */}
-              <div id="registrationTab" className="tab-pane tabs_item">
-                <HtmlContent>{event.registration?.widgetCode!}</HtmlContent>
-              </div>
-            </div>
-          </div>
-        </div>
+        <EventTabs event={event} />
       </div>
     </>
   )
@@ -279,29 +177,4 @@ const EventDetails = (props: { event: Event }) => {
   function isOpen() {
     return event.status === Enum_Event_Status.Open
   }
-
-  function concatWithoutDuplicates(
-    arr1: PlayerEntity[],
-    arr2: PlayerEntity[],
-  ): PlayerEntity[] {
-    const newArray: PlayerEntity[] = [...arr1]
-
-    arr2.map((i) => {
-      if (
-        newArray.findIndex(
-          (player) => player.attributes?.name == i.attributes?.name,
-        ) < 0
-      ) {
-        newArray.push(i)
-      }
-    })
-
-    return newArray.sort((a, b) => {
-      const name1 = a.attributes?.name!
-      const name2 = b.attributes?.name!
-      return name1.localeCompare(name2)
-    })
-  }
 }
-
-export default EventDetails
