+++ src/libs/slug-params.ts
@@ -1,0 +1,5 @@
+export interface SlugParamsProps {
+  params: {
+    slug: string
+  }
+}
